+++ src/Resources/contao/dca/tl_fernschach_spieler_konto.php
@@ -47,6 +47,12 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler_konto'] = array
 		),
 		'global_operations' => array
 		(
+			'verschiebeBuchungen' => array
+			(
+				'label'               => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto']['verschiebeBuchungen'],
+				'href'                => 'key=verschiebeBuchungen',
+				'icon'                => 'bundles/contaofernschach/images/move.png'
+			),
 			'importBuchungen' => array
 			(
 				'label'               => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto']['importBuchungen'],
@@ -146,7 +152,6 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler_konto'] = array
 		),
 		'tstamp' => array
 		(
-			'sorting'                 => true,
 			'flag'                    => 6,
 			'sql'                     => "int(10) unsigned NOT NULL default '0'"
 		),
@@ -218,7 +223,6 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler_konto'] = array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto']['sortierung'],
 			'exclude'                 => true,
-			'flag'                    => 6,
 			'inputType'               => 'text',
 			'eval'                    => array
 			(
@@ -232,6 +236,7 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler_konto'] = array
 		'kategorie' => array(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto']['kategorie'],
 			'exclude'                 => true,
+			'sorting'                 => true,
 			'filter'                  => true,
 			'inputType'               => 'select',
 			'options'                 => &$GLOBALS['TL_LANG']['tl_fernschach_spieler_konto']['kategorie_options'],
@@ -491,7 +496,7 @@ class tl_fernschach_spieler_konto extends \Backend
 		if($arrRow['markierung']) $css .= 'background-color:#FFE8DD; ';
 
 		$args[0] = '<span style="'.$css.'">'.$args[0].'</span>';
-		$args[1] = '<span style="'.$css.'">'.$args[1].'</span>';
+		$args[1] = '<span style="'.$css.'">'.($args[1] ? $args[1] : '-').'</span>';
 		$args[2] = '<span style="white-space:nowrap;'.$css.'">'.self::getEuro($args[2], $arrRow['typ']).'</span>';
 		$args[3] = '<span style="'.$css.'">'.$args[3].'&nbsp;</span>';
 		$args[4] = '<span style="white-space:nowrap;'.$css.'">'.$args[4].'&nbsp;</span>';